formatting
Added a new font, renamed the css for my index page, created another css to start making my own stylization for the hover menu on the front page. Still working on what colors and stuff I want. Trying to replicate a retro windows 95 style, hence the acquisition of the font.